requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.38.0
deprecation==2.1.0
Django==4.2.9
django-ckeditor==6.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
dnspython==2.4.2
exceptiongroup==1.2.0
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
packaging==23.2
pillow==10.2.0
postgrest==0.13.2
psycopg2==2.9.9
pydantic==2.5.3
pydantic-core==2.14.6
pymongo==4.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
realtime==1.0.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.3
supafunc==0.3.3
typing-extensions==4.9.0
urllib3==2.1.0
websockets==11.0.3
